Minor Changes ( worst commit msg ever )

+++ scheduler.c
@@ -38,8 +38,10 @@ Queue* initQueue() {
 
 int isEmpty(Queue* pQueue){ return (pQueue->size == 0); }
 
-int enqueue(Queue* pQueue, Process proc) {
-    if(pQueue->size == maxProcess || pQueue == NULL) {
+int enqueue(Queue* pQueue, Process proc) 
+{
+    if(pQueue->size == maxProcess || pQueue == NULL) 
+    {
         puts("Queue is full or empty!");
         return -1;
     }
@@ -50,8 +52,10 @@ int enqueue(Queue* pQueue, Process proc) {
     return 1;
 }
 
-Process* deque(Queue* pQueue) {
-    if(isEmpty(pQueue)) {
+Process* deque(Queue* pQueue) 
+{
+    if(isEmpty(pQueue)) 
+    {
         puts("You are trying to deque from an empty queue");
         return NULL;
     }
@@ -63,11 +67,12 @@ Process* deque(Queue* pQueue) {
     return ret;
 }
 
-Process* getQueueHead(Queue* pQueue) {
-    if(isEmpty(pQueue)) {
-        return NULL;
-    }
+Process* getQueueHead(Queue* pQueue) 
+{
+    if(isEmpty(pQueue)) return NULL;
+
     Process* ret = &((pQueue->info)[pQueue->head]);
+    
     return ret;
 }
 
@@ -88,14 +93,17 @@ Process* realTimeProc[60];
 int totalProcesses = 0;
 int finishedProcesses = 0;
 
-void initRealTimeProcesses() {
+void initRealTimeProcesses() 
+{
     // inicializando o vetor de realTime com processos "nulos"
-    for(int i = 0; i < 60; ++i) {
+    for(int i = 0; i < 60; ++i) 
+    {
         realTimeProc[i] = NULL;
     }
 }
 
-void initProcessesQueues() {
+void initProcessesQueues() 
+{
     // priority queues
     for(int i = 0; i < numPriorities; ++i) priorityProc[i] = initQueue();
     
@@ -104,6 +112,7 @@ void initProcessesQueues() {
     
     // realTime processes
     memset(occupiedByRealTime, 0, sizeof occupiedByRealTime); // initially all time slots are free
+    
     initRealTimeProcesses();
 }
 
@@ -111,54 +120,66 @@ void initProcessesQueues() {
 /*
  INSERT NEW PROCESS INTO RESPECTIVE QUEUE
  */
-void newPriority(Process* p) {
+void newPriority(Process* p) 
+{
     enqueue(priorityProc[p->priority], *p);
 }
 
-void newRoundRobin(Process* p) {
+void newRoundRobin(Process* p) 
+{
     enqueue(roundRobinProc, *p);
 }
 
-void newRealTime(Process* p) {
-    if(p->start + p->duration > 60) {
+void newRealTime(Process* p) 
+{
+    if(p->start + p->duration > 60) 
+    {
         printf("Erro de overflow na criacao do processo de nome = %s\n", p->name);
         exit(-1);
     }
-    for(int i = p->start; i < p->start + p->duration; ++i) {
-        if(occupiedByRealTime[i]) {
+    for(int i = p->start; i < p->start + p->duration; ++i) 
+    {
+        if(occupiedByRealTime[i]) 
+        {
             printf("Erro, processo de nome %s quer usar segundos ocupados\n", p->name);
             exit(-2);
         }
     }
-
     realTimeProc[p->start] = p;
+    
     printf("realTimeProc[start - %d]: duration - %d an finished\n", realTimeProc[p->start]->start, realTimeProc[p->start]->duration);
     for(int i = p->start; i < p->start + p->duration; ++i) occupiedByRealTime[i] = 1;
     // inseri um processo realTime
 }
 
 // Get next process to execute
-void dequeueNextProcess() {
-    for(int i = 0; i < numPriorities; ++i) {
-        if( !isEmpty(priorityProc[i]) ) {
+void dequeueNextProcess() 
+{
+    for(int i = 0; i < numPriorities; ++i) 
+    {
+        if( !isEmpty(priorityProc[i]) ) 
+        {
             printf("dequed priority\n");
             deque(priorityProc[i]);
         }
     }
-    if( !isEmpty(roundRobinProc) )  {
+    if( !isEmpty(roundRobinProc) )  
+    {
         printf("dequed round robin\n");
         deque(roundRobinProc);
     }
 }
 
 // Get next process to execute
-Process* getNextProcess() {
-    for(int i = 0; i < numPriorities; ++i) {
-        if( !isEmpty(priorityProc[i]) ) {
-            return getQueueHead(priorityProc[i]);
-        }
+Process* getNextProcess() 
+{
+    for(int i = 0; i < numPriorities; ++i) 
+    {
+        if( !isEmpty(priorityProc[i]) ) return getQueueHead(priorityProc[i]);
     }
-    if( !isEmpty(roundRobinProc) )  {
+
+    if( !isEmpty(roundRobinProc) )  
+    {
         return getQueueHead(roundRobinProc);
     }
     
@@ -175,35 +196,40 @@ void newProcessHandler(int signal)
     
     printf("Receive program %s process of type %d\n", programName, params[0]);
     
-    if(params[0] == -1) {
-        return; // condicao de termino
-    }
+    if(params[0] == -1) return;
     totalProcesses++;
 
     Process* p = (Process*) malloc( sizeof ( Process ) );
 
     strcpy(p->name, programName);
+    
     p->type = params[0];
     
     // Initializes process and stops execution in order to obtain pid
     pid_t pid = fork();
-    if(pid != 0) {
+    if(pid != 0) 
+    {
         p->procPid = pid;
         kill(pid, SIGSTOP);
-    } else {
-        execve(programName, NULL, NULL);
-    }
-    if(pid != 0) { 
+    } 
+    else execve(programName, NULL, NULL);
+    
+
+    if(pid != 0) 
+    { 
         p->finished = 0;
-        if(params[0] == 1) {
+        if(params[0] == 1) 
+        {
             p->priority = -1;
             newRoundRobin(p);
         }
-        else if(params[0] == 2) {
+        else if(params[0] == 2)
+        {
             p->priority = params[1];
             newPriority(p);
         }
-        else if(params[0] == 3) {
+        else if(params[0] == 3) 
+        {
             p->start = params[1];
             p->duration = params[2];
             newRealTime(p);
@@ -215,26 +241,47 @@ void newProcessHandler(int signal)
 
 }
 
-void enqueueInterruptedProcess(Process* interruptedProcess) {
+void enqueueInterruptedProcess(Process* interruptedProcess) 
+{
     int processPriority = interruptedProcess->priority;
-    if (processPriority == -1) {
+    if (processPriority == -1) 
+    {
         printf("enqueued round robin\n");
         enqueue(roundRobinProc, *interruptedProcess);
-    } else {
+    } 
+    else 
+    {
         printf("enqueued priority\n");
         enqueue(priorityProc[processPriority], *interruptedProcess);
     }
 }
 
-Process* switchProcesses(Process* curProcess, Process* nProcess) {
+Process* switchProcesses(Process* curProcess, Process* nProcess) 
+{
     printf("\nstopping process %s and starting %s\n", curProcess->name, nProcess->name);
-    enqueueInterruptedProcess(curProcess);
-    if(curProcess != NULL) kill(curProcess->procPid, SIGSTOP); // stop current process
+    
+    // so colocar de volta na fila, se nao for realTime obviamente
+    if(curProcess->type != 3) enqueueInterruptedProcess(curProcess);
+    
+    if(curProcess != NULL)
+    {
+    	printf("Vou matar %s\n", curProcess->name);
+
+    	kill(curProcess->procPid, SIGSTOP); // stop current process
+    	
+    	printf("BIZUU %d %d\n", curProcess->procPid, nProcess->procPid);
+    }
+    printf("BIZUU %d %d\n", curProcess->procPid, nProcess->procPid);
+
+    printf("Vou continuar %s\n", nProcess->name);
+
     kill(nProcess->procPid, SIGCONT); // start next process
+    
     return nProcess;
 }
 
-void scheduler() {
+void scheduler() 
+{
     int scheduler_mem = shmget(202, sizeof(pid_t), IPC_CREAT | S_IRUSR | S_IRWXU );
     pid_t* scheduler_pid = (int*) shmat(scheduler_mem, 0, 0);
     *scheduler_pid = getpid();
@@ -250,14 +297,18 @@ void scheduler() {
     {
         printf("error registering signal\n");
     }
+
     printf("registered signal handler\n");
    
     initProcessesQueues();
+
     clock_t stTime, curTime;
+    
     stTime = clock();
-    int currentSecond = 0;
-    int prevSecond = 0;
+    
+    int currentSecond = 0, prevSecond = 0;
     Process* curProcess = NULL;
+
     int changedProcess = 0;
     int result, status; // temporary variables to store process meta-data
     
@@ -267,22 +318,30 @@ void scheduler() {
     // 1 if a round robin process is in execution, 0 otherwise
     int executingRoundRobinProcess = 0;
     clock_t rrStartTime;
-    while(totalProcesses == 0 || totalProcesses != finishedProcesses) {
+
+    while(totalProcesses == 0 || totalProcesses != finishedProcesses) 
+    {
         curTime = clock();
         currentSecond = ((curTime - stTime) / CLOCKS_PER_SEC) % 60;
-        if (executingRealTimeProcess) {
+        
+        if (executingRealTimeProcess) 
+        {
             result = waitpid(curProcess->procPid, &status, WNOHANG);
             // printf("real time process - result: %d\n\n", result);
-            if (result != 0) { // process finished execution
+            if (result != 0) 
+            { // process finished execution
                 printf("Process %s finished\n", curProcess->name);
                 curProcess->finished = 1;
                 realTimeProc[curProcess->start] = NULL;
                 finishedProcesses++;
                 executingRealTimeProcess = 0;
-            } else {
+            } 
+            else 
+            {
                 // if a real time process is executing we must check if its time
                 // is up and kill it if necessary
-                if (curProcess->duration + curProcess->start == currentSecond) {
+                if (curProcess->type == 3 && curProcess->duration + curProcess->start == currentSecond) 
+                {
                     kill(curProcess->procPid, SIGSTOP);
                     executingRealTimeProcess = 0;
                 }
@@ -291,33 +350,45 @@ void scheduler() {
         
         // new real time process starting at the current second
         // that has not yet finished
-        if (currentSecond != prevSecond) {
+        if (currentSecond != prevSecond) 
+        {
             printf("current second %d\n", currentSecond);
-            if (realTimeProc[currentSecond] != NULL) {
+            if (realTimeProc[currentSecond] != NULL) 
+            {
                 printf("start = %d\nduration=%d\nfinished = %d\n\n", realTimeProc[currentSecond]->start, realTimeProc[currentSecond]->duration, realTimeProc[currentSecond]->finished);
             }
             prevSecond = currentSecond;
         }
-        if(!executingRealTimeProcess &&
+        if( !executingRealTimeProcess &&
             realTimeProc[currentSecond] != NULL &&
-            !realTimeProc[currentSecond]->finished) {
+            ! (realTimeProc[currentSecond]->finished) ) 
+        {
+        	printf("Vamos botar %s pra rodar nessa porra", realTimeProc[currentSecond]->name);
+            printf("processoAtual = %s, next = %s\n", curProcess->name, realTimeProc[currentSecond]->name);
+            printf("Valor de curProcess antes = %s\n", curProcess->name);
             curProcess = switchProcesses(curProcess, realTimeProc[currentSecond]);
+            printf("Valor de curProcess depois = %s\n", curProcess->name);
             changedProcess = 1;
             executingRealTimeProcess = 1;
-            printf("executing real time process\n\n");
+            //printf("executing real time process\n\n");
         }
         
         // If no real time process is currently in execution
         // check which process to execute next. Next process will
-        if (!executingRealTimeProcess) {
+        if (!executingRealTimeProcess) 
+        {
             Process* nProcess = getNextProcess();
-            if (nProcess != NULL) { // there is an enqueued process to be executed
-                if(curProcess != NULL) {
+            if (nProcess != NULL) 
+            { // there is an enqueued process to be executed
+                if(curProcess != NULL) 
+                {
                     // if a round robin process is executing we must check if its time
                     // is up before switching processes in case the next process has
                     // equal or less priority (in our case, if it is a round robin process)
-                    if (nProcess->priority == -1) {
-                        if (executingRoundRobinProcess) {
+                    if (nProcess->priority == -1) 
+                    {
+                        if (executingRoundRobinProcess) 
+                        {
                             // if the current process has surpassed its quantum, then it should
                             // be switched
                             int clocks_for_quantum = CLOCKS_PER_SEC * quantum;
@@ -329,44 +400,56 @@ void scheduler() {
                                 rrStartTime = curTime;
                             }
                         }
-                    } else { // next process to be executed is priority
+                    } 
+                    else 
+                    { // next process to be executed is priority
                         // stop current process if priority is greater than current executing process
                         // printf("priority next %d, cur priority: %d\n", nProcess->priority, curProcess->priority);
-                        if (nProcess->priority > curProcess->priority) {
+                        if (nProcess->priority > curProcess->priority) 
+                        {
                             dequeueNextProcess();
                             curProcess = switchProcesses(curProcess, nProcess);
                             executingRoundRobinProcess = 0;
                             changedProcess = 1;
                         }
                     }
-                } else {
+                } 
+                else 
+                {
                     dequeueNextProcess();
                     curProcess = nProcess;
                     printf("init cur process, %s\n", curProcess->name);
                     changedProcess = 1;
-                    if (curProcess->priority == -1) { // save start time for round robin processes
+                    if (curProcess->priority == -1) 
+                    { // save start time for round robin processes
                         executingRoundRobinProcess = 1;
                         rrStartTime = curTime;
                     }
                 }
             }
 
-            if (curProcess != NULL) {
+            if (curProcess != NULL) 
+            {
                 result = waitpid(curProcess->procPid, &status, WNOHANG);
                 // printf("name: %s, result: %d\n", curProcess->name, result);
-                if (result != 0 && !curProcess->finished) {
-                    if (executingRoundRobinProcess) {
+                if (result != 0 && !curProcess->finished) 
+                {
+                    if (executingRoundRobinProcess) 
+                    {
                         executingRoundRobinProcess = 0;
                     }
                     printf("Process %s finished\n", curProcess->name);
                     curProcess->finished = 1;
                     finishedProcesses++;
-                } else {
+                } 
+                else 
+                {
                     kill(curProcess->procPid, SIGCONT); // resume process
                 }
             }
         }
-        if (curProcess != NULL && changedProcess) {
+        if (curProcess != NULL && changedProcess) 
+        {
             printf("Executing process %s\n", curProcess->name);
             changedProcess = 0;
         }
@@ -376,7 +459,8 @@ void scheduler() {
     shmctl(scheduler_mem, IPC_RMID, 0);
 }
 
-int main() {
+int main() 
+{
     scheduler();
     return 0;
 }
